requirements.txt
boto3==1.26.137
bcrypt==4.0.1
PyJWT==2.7.0
